requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.4
celery==5.6.2
celery-types==0.26.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
crispy-bootstrap4==2024.10
Django==5.1.2
django-crispy-forms==2.3
kombu==5.6.2
Markdown==3.7
packaging==26.0
pillow==11.0.0
prompt_toolkit==3.0.52
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
redis==7.3.0
six==1.17.0
sqlparse==0.5.1
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
